+++ CHANGES.md
@@ -19,6 +19,10 @@
   over the module list
   ([PR #322](https://github.com/cloudpipe/cloudpickle/pull/322)).
 
+- Add support for out-of-band pickling (Python 3.8 and later).
+  https://docs.python.org/3/library/pickle.html#example
+  ([issue #308](https://github.com/cloudpipe/cloudpickle/pull/308))
+
 1.2.2
 =====